+++ src/unpoly/fragment.js
@@ -13,7 +13,7 @@ A fragment is an element with some additional properties that are useful in the 
 a server-rendered web application:
 
 - Fragments are [identified by a CSS selector](/target-derivation), like a `.class` or `#id`.
-- Fragments are usually updated by a [link](/a-up-follow) for [form](/form-up-submit) that targets their selector.
+- Fragments are usually updated by a [link](/a-up-follow) for [form](/up-submit) that targets their selector.
   When the server renders HTML with a matching element, the fragment is swapped with a new version.
 - As fragments enter the page they are automatically [compiled](/up.compiler) to activate [JavaScript behavior](/up.script).
 - Fragment changes may be [animated](/up.motion).